# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "GreenLync Event-Driven API Gateway Boilerplate",
        "title": "GreenLync API Gateway",
        "termsOfService": "https://greenlync.com",
        "contact": {
            "name": "API Support",
            "email": "zeelrupapara@gmail.com"
        },
        "license": {
            "name": "GreenLync",
            "url": "https://greenlync.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8888",
    "basePath": "/",
    "paths": {
...
        "/api/v1/emails": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get All InEmails",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Emails"
                ],
                "operationId": "GetAllInEmails",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Mail"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.HttpResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create InEmail",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Emails"
                ],
                "operationId": "CreateInEmail",
                "parameters": [
                    {
                        "description": "InEmail Request Body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.InEmail"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/v1.InEmail"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.HttpResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/emails/me/bin": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get Account InEmails Bin",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Emails"
                ],
                "operationId": "GetAccountBinEmails",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Mail"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.HttpResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/emails/me/draft": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get Account InEmails Draft",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Emails"
                ],
                "operationId": "GetAccountDraftEmails",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Mail"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.HttpResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/emails/me/inbox": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get Account InEmails Inbox",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Emails"
                ],
                "operationId": "GetAllAccountInEmails",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Mail"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.HttpResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/emails/me/outbox": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get Account InEmails Mail",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Emails"
                ],
                "operationId": "GetAllAccountOutEmails",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Mail"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.HttpResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/emails/me/{tracking_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get Account InEmail Details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Emails"
                ],
                "operationId": "GetAccountInEmail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tracking ID",
                        "name": "tracking_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Mail"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.HttpResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/emails/test": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Test SMTP",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Emails"
                ],
                "operationId": "TestSMTPEmail",
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.HttpResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "http.HttpResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Http status Code",
                    "type": "integer"
                },
                "data": {
                    "description": "if the request were successful the data will be saved here"
                },
                "error": {
                    "description": "Generic General Error Message defined in the system",
                    "type": "string"
                },
                "message": {
                    "description": "More detailed error message indicates why the request was unsuccessful",
                    "type": "string"
                },
                "success": {
                    "description": "Response flag indicates whether the HTTP request was successful or not",
                    "type": "boolean"
                }
            }
        },
...
        "model.Mail": {
            "type": "object",
            "properties": {
                "body": {
                    "description": "smtp \u0026 in-email \u0026 chat",
                    "type": "string"
                },
                "common_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted": {
                    "type": "boolean"
                },
                "edited": {
                    "type": "boolean"
                },
                "email_tracking_id": {
                    "type": "string"
                },
                "from": {
                    "description": "smtp",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "original": {
                    "description": "in the admin panel the broker only need to see one version of the email",
                    "type": "boolean"
                },
                "owner_id": {
                    "type": "integer"
                },
                "read_at": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/model.MailStatus"
                },
                "subject": {
                    "description": "smtp \u0026 in-email",
                    "type": "string"
                },
                "to": {
                    "description": "smtp",
                    "type": "string"
                },
                "to_user_id": {
                    "description": "in-email \u0026 chat",
                    "type": "integer"
                },
                "type": {
                    "$ref": "#/definitions/model.MailType"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/model.User"
                },
                "user_id": {
                    "description": "in-email \u0026 chat",
                    "type": "integer"
                }
            }
        },
        "model.MailStatus": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                4
            ],
            "x-enum-varnames": [
                "MailStatus_queue",
                "MailStatus_sent",
                "MailStatus_pending",
                "MailStatus_draft",
                "MailStatus_dropped"
            ]
        },
        "model.MailType": {
            "type": "integer",
            "enum": [
                0,
                1,
                2
            ],
            "x-enum-varnames": [
                "MailType_smtp",
                "MailType_notification",
                "MailType_system"
            ]
        },
...
        "v1.InEmail": {
            "type": "object",
            "required": [
                "body",
                "status",
                "subject",
                "to"
            ],
            "properties": {
                "body": {
                    "type": "string"
                },
                "reply_to": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/model.MailStatus"
                },
                "subject": {
                    "type": "string"
                },
                "to": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        },
        "BearerAuth": {
            "description": "Type \"BearerAuth\" followed by a space and Access token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}